Debugging EG coding.

+++ dbs/eg_coding.py
@@ -4,11 +4,11 @@ Exposes Python bindings.
 
 Functions:
 encode(x) -> y:
-    x: 1D int32 tensor.
-    y: 1D uint8 tensor.
+    x: 1D int64 tensor.
+    y: 1D uint64 tensor.
 decode(y) -> x:
-    y: 1D uint8 tensor.
-    x: 1D int32 tensor.
+    y: 1D uint64 tensor.
+    x: 1D int64 tensor.
 """
 
 import torch
@@ -24,4 +24,4 @@ encode = _eg_coding.encode
 decode = _eg_coding.decode
 
 # Warm up
-encode(torch.randint(-1000, 1000, (10,), dtype=torch.int32))
+encode(torch.randint(-1000, 1000, (10,), dtype=torch.int64))
